initial commit -- working prototype
I had to make several compromises, and the code is not 100% good, but it is working!

I would prefer to have syntax be whitespace-sensitive, so one could write

  on &cell $do =
    to &cell
    $do
    from &cell

but { braces } are just easier to parse. They'll do for now.

Also, the language has no semantic checks right now. For instance, an &offset argument into a $brainfuck parameter won't be prevented.

So, I'm far from happy with the product as it is. But it's a start!